Moved vault command to own class and slightly cleaned it up
Took 9 minutes
